Fixed: Changed type of price field: double to BigDecimal

+++ src/main/java/com/alexshuvaev/topjava/gp/domain/Dish.java
@@ -1,6 +1,7 @@
 package com.alexshuvaev.topjava.gp.domain;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Objects;
 
@@ -12,7 +13,7 @@ import java.util.Objects;
 @Table(name = "dish", uniqueConstraints = {@UniqueConstraint(columnNames = {"date"}, name = "dish_unique_date_idx")})
 public class Dish extends AbstractNamedEntity {
     @Column(name = "price")
-    private Double price;
+    private BigDecimal price;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "restaurant_id", nullable = false)
@@ -24,14 +25,14 @@ public class Dish extends AbstractNamedEntity {
     protected Dish() {
     }
 
-    public Dish(Integer id, String name, Double price, Restaurant restaurant, LocalDate localDate) {
+    public Dish(Integer id, String name, BigDecimal price, Restaurant restaurant, LocalDate localDate) {
         super(id, name);
         this.price = price;
         this.restaurant = restaurant;
         this.date = localDate;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
